backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.7
whitenoise==6.6.0
gunicorn==21.2.0
google-generativeai==0.3.2
requests==2.31.0
django-redis==5.4.0
celery==5.3.4
redis==5.0.1
drf-spectacular==0.26.5
drf-nested-routers==0.93.4
